+++ lib/src/widgets/chat.dart
@@ -50,6 +50,7 @@ class Chat extends StatefulWidget {
     this.onPreviewDataFetched,
     required this.onSendPressed,
     this.onTextChanged,
+    this.onTextFieldTap,
     this.sendButtonVisibilityMode = SendButtonVisibilityMode.editing,
     this.showUserAvatars = false,
     this.showUserNames = false,
@@ -146,6 +147,9 @@ class Chat extends StatefulWidget {
   final Widget? inputSuffixIcon;
 
   final void Function()? onTapInput;
+
+  /// See [Input.onTextFieldTap]
+  final void Function()? onTextFieldTap;
 
   /// See [Input.sendButtonVisibilityMode]
   final SendButtonVisibilityMode sendButtonVisibilityMode;
@@ -281,8 +285,8 @@ class _ChatState extends State<Chat> {
   ) {
     return Center(
       child: SizedBox(
-        width: 20.0,
-        height: 20.0,
+        width: 20,
+        height: 20,
         child: CircularProgressIndicator(
           value: event == null || event.expectedTotalBytes == null
               ? 0
@@ -418,6 +422,7 @@ class _ChatState extends State<Chat> {
                           inputHeader: widget.inputHeader,
                           inputSuffixIcon: widget.inputSuffixIcon,
                           disableInput: widget.disableInput,
+                          onTextFieldTap: widget.onTextFieldTap,
                           sendButtonVisibilityMode:
                               widget.sendButtonVisibilityMode,
                         ),