+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.4.0
+
+- Update to Flutter 2.5
+
+## 1.3.4
+
+- Update dependencies
+
+## 1.3.3
+
+- Add `onTextFieldTap`. Thanks @halildurmus for the PR!
+- Add `messageInsetsHorizontal` and `messageInsetsVertical` to the theme to customize message bubble's paddings
+
 ## 1.3.2
 
 - Fix memory leak. Thanks @m-j-g for reporting!